Move A* defs into astar.hpp
